feat: GUI views分割実装 - 539行→176行最大（67%削減）
Phase2 Week1: GUI系リファクタリング第1弾完了

## 📊 分割結果
- gui_views.py: 539行 → 5モジュール分離
- main_window.py: 49行（ウィンドウ管理）
- widgets.py: 163行（基本ウィジェット）
- controls.py: 176行（コントロール要素）
- dialogs.py: 93行（ダイアログ）
- main_view.py: 96行（統合ビュー）

## 🏗️ アーキテクチャ改善
- Single Responsibility Principle適用
- MVC分離強化
- 後方互換性維持（レガシーラッパー）
- TYPE_CHECKING使用で循環インポート回避

## ✅ 動作確認済み
- 新モジュール構造インポート成功
- レガシー互換性確認成功

残り36件の技術的負債へ継続対応予定
